update weeklyplan when blokingplan created and updated

+++ Com.Bateeq.Service.Masterplan.Lib/Models/WeeklyPlanItem.cs
@@ -15,10 +15,10 @@ namespace Com.Bateeq.Service.Masterplan.Lib.Models
         public int Efficiency { get; set; }
         public int Operator { get; set; }
         public int WorkingHours { get; set; }
-        public int AhTotal { get; set; }
-        public int EhTotal { get; set; }
-        public int UsedEh { get; set; }
-        public int RemainingEh { get; set; }
+        public double AhTotal { get; set; }
+        public double EhTotal { get; set; }
+        public double UsedEh { get; set; }
+        public double RemainingEh { get; set; }
         public DateTimeOffset StartDate { get; set; }
         public DateTimeOffset EndDate { get; set; }
